# openapi.yaml
openapi: 3.0.3
info:
  title: Revvo Smart Price API
  description: |
    API do sistema Revvo Smart Price — plataforma de gestão de terminais de consulta de preços e exibição de mídia em lojas.

    ## Autenticação

    A API possui dois mecanismos de autenticação:

    ### 1. Web Session (Cookie JWT)
    Utilizada pelo painel administrativo. O login via `POST /api/auth/login` retorna um cookie `revvo_session` (httpOnly, SameSite=lax) com validade de 8 horas.
    Todas as rotas administrativas exigem esse cookie.

    ### 2. Terminal Bearer Token
    Utilizada pelos dispositivos físicos (terminais). O terminal obtém um JWT via `POST /api/terminal/activate` (primeira vez) ou `POST /api/terminal/auth` (reautenticação).
    O token deve ser enviado no header `Authorization: Bearer <token>` e tem validade de 30 dias.

    ## Perfis de acesso

    - **Master** (`revvo-master`): acesso irrestrito a todas as empresas.
    - **Usuário comum**: acesso restrito à própria empresa.
  version: 1.0.0
  contact:
    name: Revvo Smart Price
servers:
  - url: https://revvo-smart-price-backend.vercel.app
    description: Produção
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: revvo_session
      description: Cookie JWT gerado pelo login web. Válido por 8 horas.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT gerado na ativação do terminal. Válido por 30 dias.
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        isActive:
          type: boolean
        companyId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        company:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
    Store:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
          nullable: true
        companyId:
          type: string
          format: uuid
        company:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        license:
          $ref: '#/components/schemas/StoreLicense'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StoreLicense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quantityTotal:
          type: integer
          example: 10
        quantityUsed:
          type: integer
          example: 3
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, expired, suspended]
        renewals:
          type: array
          items:
            $ref: '#/components/schemas/LicenseRenewal'
    LicenseRenewal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        previousExpiresAt:
          type: string
          format: date-time
        newExpiresAt:
          type: string
          format: date-time
        addedQuantity:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ean:
          type: string
          example: "7891000315507"
        name:
          type: string
          example: Leite Integral 1L
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
          example: 4.99
        offerPrice:
          type: number
          nullable: true
          example: 3.99
        clubPrice:
          type: number
          nullable: true
          example: 3.49
        stock:
          type: integer
          nullable: true
        isActive:
          type: boolean
        externalId:
          type: string
          nullable: true
        storeId:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
...
security:
  - cookieAuth: []
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Listar usuários
      description: Master retorna todos. Usuário comum retorna apenas da própria empresa.
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Users]
      summary: Criar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                isActive:
                  type: boolean
                  default: true
                companyId:
                  type: string
                  format: uuid
                  description: Apenas master pode especificar. Não-master usa a própria empresa.
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: E-mail já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Users]
      summary: Editar usuário
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                isActive:
                  type: boolean
                password:
                  type: string
                  description: Será hasheado com bcrypt
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: E-mail já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Users]
      summary: Remover usuário
      responses:
        '200':
          description: Usuário removido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ─────────────────────────────────────────────
  # PROFILE
  # ─────────────────────────────────────────────
...
  /api/stores:
    get:
      tags: [Stores]
      summary: Listar lojas
      responses:
        '200':
          description: Lista de lojas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Store'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Stores]
      summary: Criar loja
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                address:
                  type: string
                companyId:
                  type: string
                  format: uuid
                  description: Apenas master pode especificar.
      responses:
        '201':
          description: Loja criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '400':
          description: Nome é obrigatório
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/stores/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Stores]
      summary: Buscar loja por ID
      responses:
        '200':
          description: Dados da loja
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '404':
          description: Loja não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Stores]
      summary: Editar loja
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
      responses:
        '200':
          description: Loja atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
    delete:
      tags: [Stores]
      summary: Remover loja
      responses:
        '200':
          description: Loja removida
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ─────────────────────────────────────────────
  # LICENSE
  # ─────────────────────────────────────────────
...
  /api/terminal/bootstrap:
    get:
      tags: [Terminal Device - Data]
      summary: Bootstrap do terminal (dados iniciais)
      description: Retorna configuração, loja, empresa e endpoints que o terminal deve usar.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados de bootstrap
          content:
            application/json:
              schema:
                type: object
                properties:
                  terminal:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      location:
                        type: string
                      isPriceChecker:
                        type: boolean
                      isMediaDisplay:
                        type: boolean
                  store:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      address:
                        type: string
                  company:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                  config:
                    type: object
                    properties:
                      resetTime:
                        type: integer
                  endpoints:
                    type: object
                    properties:
                      productByEan:
                        type: string
                      catalog:
                        type: string
                      publicProductByEan:
                        type: string
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Terminal bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/terminal/product/{ean}:
    parameters:
      - name: ean
        in: path
        required: true
        schema:
          type: string
        example: "7891000315507"
    get:
      tags: [Terminal Device - Data]
      summary: Buscar produto por EAN
      description: O storeId é derivado do JWT do terminal.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...